+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Schema literals weren't treated as their `typeof` type, which prevented multi-type union detection.
+- `FormPath` and `FormPathLeaves` didn't narrow the types correctly for `Date`, `Set` and `File`.
 
 ## [2.6.1] - 2024-02-24